Update to m3u8-rs 6
Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2117>

+++ net/hlssink3/src/hlscmafsink/imp.rs
@@ -404,7 +404,7 @@ impl HlsCmafSink {
 
         let playlist = MediaPlaylist {
             version: Some(6),
-            target_duration: target_duration as f32,
+            target_duration: target_duration as u64,
             playlist_type,
             independent_segments: true,
             ..Default::default()
